# apps/cockpit/openapi.json
{
  "components": {
    "schemas": {
      "CheckDefOut": {
        "properties": {
          "expect": {
            "title": "Expect",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "owned_by": {
            "default": "platform",
            "title": "Owned By",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "sql": {
            "title": "Sql",
            "type": "string"
          },
          "type": {
            "default": "",
            "title": "Type",
            "type": "string"
          },
          "unit": {
            "default": "",
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "name",
          "sql",
          "expect",
          "severity"
        ],
        "title": "CheckDefOut",
        "type": "object"
      },
...
      "CompileOut": {
        "properties": {
          "checks": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CheckDefOut"
            },
            "title": "Checks",
            "type": "array"
          },
          "conflicts": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "dataset": {
            "title": "Dataset",
            "type": "string"
          },
          "determinism_hash": {
            "default": "",
            "title": "Determinism Hash",
            "type": "string"
          },
          "product": {
            "title": "Product",
            "type": "string"
          },
          "yaml_preview": {
            "default": "",
            "title": "Yaml Preview",
            "type": "string"
          }
        },
        "required": [
          "product",
          "dataset"
        ],
        "title": "CompileOut",
        "type": "object"
      },
      "ContractIn": {
        "description": "Lifecycle ist bewusst KEIN Eingabefeld \u2014 \u00dcberg\u00e4nge laufen nur \u00fcber\napprove/deprecate; PUT erzwingt draft (S-2).",
        "properties": {
          "dataset": {
            "default": "",
            "title": "Dataset",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "guarantees": {
            "additionalProperties": true,
            "default": {},
            "title": "Guarantees",
            "type": "object"
          },
          "owned_by": {
            "default": "platform",
            "title": "Owned By",
            "type": "string"
          },
          "owners": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Owners",
            "type": "array"
          },
          "product": {
            "title": "Product",
            "type": "string"
          },
          "version": {
            "default": "0.1.0",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "product"
        ],
        "title": "ContractIn",
        "type": "object"
      },
      "ContractOut": {
        "properties": {
          "compliance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compliance"
          },
          "dataset": {
            "default": "",
            "title": "Dataset",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "guarantees": {
            "additionalProperties": true,
            "default": {},
            "title": "Guarantees",
            "type": "object"
          },
          "lifecycle": {
            "default": "draft",
            "title": "Lifecycle",
            "type": "string"
          },
          "owned_by": {
            "default": "platform",
            "title": "Owned By",
            "type": "string"
          },
          "owners": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Owners",
            "type": "array"
          },
          "product": {
            "title": "Product",
            "type": "string"
          },
          "version": {
            "default": "0.1.0",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "product"
        ],
        "title": "ContractOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "DQ & Observability Cockpit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/checks/{dataset}/dry-run": {
      "post": {
        "description": "Compile contract for dataset and run checks without persisting results.\n\nBody: { \"environment\": \"<env-name>\", \"execution_mode\": \"auto|batch|isolated\" }\nOhne Environment: compile-only preview. Ergebnisse werden NIE persistiert.",
        "operationId": "dry_run_checks_api_checks__dataset__dry_run_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "default": {},
                "title": "Body",
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dry Run Checks",
        "tags": [
          "checks"
        ]
      }
    },
    "/api/checks/{dataset}/revert": {
      "post": {
        "description": "[AUTHZ] Revert checks/{dataset}/checks.yml to the previous git version (F7).",
        "operationId": "revert_checks_api_checks__dataset__revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset",
            "required": true,
            "schema": {
              "title": "Dataset",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revert Checks",
        "tags": [
          "checks"
        ]
      }
    },
...
    "/api/contracts/{product}": {
      "get": {
        "operationId": "get_contract_api_contracts__product__get",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Contract",
        "tags": [
          "contracts"
        ]
      },
      "put": {
        "description": "[CONTRACT-SQL-FREE] [AUTHZ] Update a contract \u2014 Gate G1 enforced here.\n\nLifecycle ist kein Eingabefeld: das Ergebnis eines PUT ist IMMER ein Draft.\nEin PUT auf einen aktiven Contract erzeugt damit ein Draft-Amendment \u2014 die\nzertifizierte Version bleibt als `.active.yml`-Snapshot die G3-Diff-Basis,\nbis ein erneutes Approve sie ersetzt.",
        "operationId": "update_contract_api_contracts__product__put",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Contract",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{product}/approve": {
      "post": {
        "description": "[AUTHZ] Promote a draft contract to active (WS2-6 / M2).\n\nServer-side breaking guard (G3): if the change is breaking versus the last\ncertified version, a SemVer **major** bump is required.",
        "operationId": "approve_contract_api_contracts__product__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Contract",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{product}/compile": {
      "post": {
        "description": "[DETERMINISM] [SCHEMA-MAP] Compile contract \u2192 CheckDefs.\n\nG1: der Contract wird auch auf dem Disk-Pfad validiert (kein Schmuggel an\nPUT vorbei). G2: '{schema}' bleibt als Platzhalter im Output. Non-dry-run\nerfordert lifecycle=active (WS3-2) und Schreibrecht.",
        "operationId": "compile_contract_api_contracts__product__compile_post",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": false,
              "title": "Dry Run",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compile Contract",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{product}/deprecate": {
      "post": {
        "description": "[AUTHZ] Retire an active contract (WS2-6).",
        "operationId": "deprecate_contract_api_contracts__product__deprecate_post",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deprecate Contract",
        "tags": [
          "contracts"
        ]
      }
    },
...
    "/api/contracts/{product}/export/bdc": {
      "post": {
        "description": "Generate CSN + ORD artifact fragments for BDC integration. (WS5-4)\n\nEinseitig (E1): files are returned as JSON for manual deployment.\nNever writes to DB or catalog.",
        "operationId": "export_bdc_api_contracts__product__export_bdc_post",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Bdc",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{product}/export/odcs": {
      "get": {
        "description": "R5-1: deterministic ODCS 3.1 export. [A1] compliance stays out.",
        "operationId": "export_odcs_api_contracts__product__export_odcs_get",
        "parameters": [
          {
            "in": "path",
            "name": "product",
            "required": true,
            "schema": {
              "title": "Product",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Odcs",
        "tags": [
          "contracts"
        ]
      }
    },
...
  }
}